"Add view function v1 and corresponding URL in mysite/main/views.py and mysite/mysite/urls.py"
The diff shows the addition of a new view function `v1` in the file `mysite/main/views.py` which returns an HTTP response with the string "<h1>view 1!</h1>". A new URL pattern is also added in the file `mysite/mysite/urls.py` to map the root URL to the new view function `v1`. The existing view function `index` in `mysite/main/views.py` is modified to return an HTTP response with the string "<h1>Man Njoro</h1>".