"Added handling for POST request in index view"
This commit adds a few lines to the index view function in the views.py file. Now, if the request method is POST, it checks if the "save" button was clicked in the form. Currently, there is no functionality associated with the "save" button, so this code just passes when the "save" button is clicked. This change allows for the addition of functionality when the form is submitted in the future.