"Added LOGIN\_REDIRECT\_URL setting to mysite/settings.py"
Explanation:
The code diff shows the addition of one line to the file mysite/settings.py in a Django project. The line added is:
LOGIN\_REDIRECT\_URL = "/"

This setting specifies the URL to redirect to after a successful login. In this case, it is set to the root URL of the site, "/". This is a required setting in Django and should be added to the settings file if it is missing. The commit message should be concise and descriptive of the change made, without making any assumptions or interpretations about why the change was made.